refactor: screens/home 부분 import 구문 단축 url 사용하는 방식으로 변경
