+++ src/components/common/layout/index.tsx
@@ -1,0 +1,4 @@
+export { default as FooterLayout } from './FooterLayout';
+export { default as InnerLayout } from './InnerLayout';
+export { default as KeyboardLayout } from './KeyboardLayout';
+export { default as Layout } from './Layout';
